refactor(map): extract MapContainer Leaflet effects into hooks (CodeHealth B->A)
Greens the last yellow file. MapContainer's 7 useEffects are moved verbatim into
four focused, documented hooks — behavior-preserving (same deps, cleanups, and
imperative Leaflet calls); the component keeps the shared refs + popup builders
and just composes them:
- useLeafletMap        — StrictMode-safe init + recenter
- usePoiMarkers        — incremental POI markers + auto-open current popup
- useUserLocationMarker — draggable user marker create + position sync
- useMapPopupNavigation — popup closer/farther/directions delegation
Plus mapTypes.ts (shared POILocation, avoids a component<->hook type cycle).

MapContainer MI 16.7 (yellow) -> green; CodeHealth 89.8 -> 92.7 (B -> A);
0 files >500 LOC; 0 circular imports. lint + type-check + full vitest suite +
vite build all green. ⚠️ Touches the core map UI — please visually validate the
Vercel preview (map renders, markers, popups, closer/farther nav, user-marker drag).
