+++ code-health/codehealth-history.tsv
@@ -1,2 +1,2 @@
 date	score	grade	documentation	maintainability	structure	resilience_worst_file_	type_size_safety	security_deps_
-2026-06-30	89.8	B	93	94.4	100	58.5	78.2	100
+2026-06-30	92.7	A	93.2	97.4	100	67	86.7	100
