+++ code-health/codehealth-stamp.json
@@ -1,16 +1,16 @@
 {
-  "badge": "B · 89.8 / 100",
-  "chart": "```text\n                            weight  score (0–100)\nDocumentation         20%   ███████████████████░    93\nMaintainability       25%   ███████████████████░  94.4   57🟢 / 2🟡\nStructure             20%   ████████████████████   100\nResilience (worst)    10%   ████████████░░░░░░░░  58.5   worst file MI 16.7\nType & size           15%   ████████████████░░░░  78.2\nSecurity (deps)       10%   ████████████████████   100   no advisories\n                            ────────────────────\nCodeHealth                  ██████████████████░░  89.8   grade B\n```",
-  "pie": "```mermaid\n%%{init: {\"theme\": \"base\", \"themeVariables\": {\"pie1\": \"#2e7d32\", \"pie2\": \"#f9a825\", \"pie3\": \"#c62828\"}}}%%\npie showData title Maintainability Index bands (59 files)\n  \"Green (>=20)\" : 57\n  \"Yellow (10-19)\" : 2\n```",
-  "files": 59,
-  "loc": "10.2k",
-  "green": 57,
-  "yellow": 2,
+  "badge": "A · 92.7 / 100",
+  "chart": "```text\n                            weight  score (0–100)\nDocumentation         20%   ███████████████████░  93.2\nMaintainability       25%   ███████████████████░  97.4   63🟢 / 1🟡\nStructure             20%   ████████████████████   100\nResilience (worst)    10%   █████████████░░░░░░░    67   worst file MI 18.4\nType & size           15%   █████████████████░░░  86.7\nSecurity (deps)       10%   ████████████████████   100   no advisories\n                            ────────────────────\nCodeHealth                  ███████████████████░  92.7   grade A\n```",
+  "pie": "```mermaid\n%%{init: {\"theme\": \"base\", \"themeVariables\": {\"pie1\": \"#2e7d32\", \"pie2\": \"#f9a825\", \"pie3\": \"#c62828\"}}}%%\npie showData title Maintainability Index bands (64 files)\n  \"Green (>=20)\" : 63\n  \"Yellow (10-19)\" : 1\n```",
+  "files": 64,
+  "loc": "10.4k",
+  "green": 63,
+  "yellow": 1,
   "red": 0,
   "mi_mean": 38,
-  "hotspots": 14,
-  "top_hotspot": "MapContainer.tsx",
-  "hotspot_table": "| Score | Revisions | Cyclomatic | File |\n|--:|--:|--:|---|\n| 1036 | 14 | 74 | [`apps/web/src/components/MapContainer.tsx`](https://github.com/PrairieAster-Ai/nearest-nice-weather/blob/main/apps/web/src/components/MapContainer.tsx) |\n| 840 | 60 | 14 | [`apps/web/src/App.tsx`](https://github.com/PrairieAster-Ai/nearest-nice-weather/blob/main/apps/web/src/App.tsx) |\n| 300 | 12 | 25 | [`apps/web/src/hooks/usePOINavigation.ts`](https://github.com/PrairieAster-Ai/nearest-nice-weather/blob/main/apps/web/src/hooks/usePOINavigation.ts) |\n| 272 | 8 | 34 | [`apps/web/src/components/ads/AdManager.tsx`](https://github.com/PrairieAster-Ai/nearest-nice-weather/blob/main/apps/web/src/components/ads/AdManager.tsx) |\n| 210 | 5 | 42 | [`apps/web/src/services/UserLocationEstimator.ts`](https://github.com/PrairieAster-Ai/nearest-nice-weather/blob/main/apps/web/src/services/UserLocationEstimator.ts) |",
+  "hotspots": 13,
+  "top_hotspot": "App.tsx",
+  "hotspot_table": "| Score | Revisions | Cyclomatic | File |\n|--:|--:|--:|---|\n| 840 | 60 | 14 | [`apps/web/src/App.tsx`](https://github.com/PrairieAster-Ai/nearest-nice-weather/blob/main/apps/web/src/App.tsx) |\n| 300 | 12 | 25 | [`apps/web/src/hooks/usePOINavigation.ts`](https://github.com/PrairieAster-Ai/nearest-nice-weather/blob/main/apps/web/src/hooks/usePOINavigation.ts) |\n| 272 | 8 | 34 | [`apps/web/src/components/ads/AdManager.tsx`](https://github.com/PrairieAster-Ai/nearest-nice-weather/blob/main/apps/web/src/components/ads/AdManager.tsx) |\n| 210 | 5 | 42 | [`apps/web/src/services/UserLocationEstimator.ts`](https://github.com/PrairieAster-Ai/nearest-nice-weather/blob/main/apps/web/src/services/UserLocationEstimator.ts) |\n| 200 | 4 | 50 | [`apps/web/src/utils/locationEstimationUtils.ts`](https://github.com/PrairieAster-Ai/nearest-nice-weather/blob/main/apps/web/src/utils/locationEstimationUtils.ts) |",
   "fanout": 13,
   "pairs": 14,
   "cross_layer": 0,
