+++ code-health/hotspot-table.md
@@ -1,7 +1,7 @@
 | Score | Revisions | Cyclomatic | File |
 |--:|--:|--:|---|
-| 1036 | 14 | 74 | [`apps/web/src/components/MapContainer.tsx`](https://github.com/PrairieAster-Ai/nearest-nice-weather/blob/main/apps/web/src/components/MapContainer.tsx) |
 | 840 | 60 | 14 | [`apps/web/src/App.tsx`](https://github.com/PrairieAster-Ai/nearest-nice-weather/blob/main/apps/web/src/App.tsx) |
 | 300 | 12 | 25 | [`apps/web/src/hooks/usePOINavigation.ts`](https://github.com/PrairieAster-Ai/nearest-nice-weather/blob/main/apps/web/src/hooks/usePOINavigation.ts) |
 | 272 | 8 | 34 | [`apps/web/src/components/ads/AdManager.tsx`](https://github.com/PrairieAster-Ai/nearest-nice-weather/blob/main/apps/web/src/components/ads/AdManager.tsx) |
 | 210 | 5 | 42 | [`apps/web/src/services/UserLocationEstimator.ts`](https://github.com/PrairieAster-Ai/nearest-nice-weather/blob/main/apps/web/src/services/UserLocationEstimator.ts) |
+| 200 | 4 | 50 | [`apps/web/src/utils/locationEstimationUtils.ts`](https://github.com/PrairieAster-Ai/nearest-nice-weather/blob/main/apps/web/src/utils/locationEstimationUtils.ts) |
