+++ code-health/maintainability-history.tsv
@@ -1,2 +1,2 @@
 date	files	mean_mi	median_mi	min_mi	green	yellow	red
-2026-06-30	59	37.9	36.4	16.7	57	2	0
+2026-06-30	64	38.3	37.5	18.4	63	1	0
